requirements.txt
-e
aws-cdk.assets==1.110.0
aws-cdk.aws-apigateway==1.110.0
aws-cdk.aws-applicationautoscaling==1.110.0
aws-cdk.aws-autoscaling==1.110.0
aws-cdk.aws-autoscaling-common==1.110.0
aws-cdk.aws-autoscaling-hooktargets==1.110.0
aws-cdk.aws-certificatemanager==1.110.0
aws-cdk.aws-cloudformation==1.110.0
aws-cdk.aws-cloudfront==1.110.0
aws-cdk.aws-cloudwatch==1.110.0
aws-cdk.aws-codebuild==1.110.0
aws-cdk.aws-codecommit==1.110.0
aws-cdk.aws-codedeploy==1.110.0
aws-cdk.aws-codeguruprofiler==1.110.0
aws-cdk.aws-codepipeline==1.110.0
aws-cdk.aws-codepipeline-actions==1.110.0
aws-cdk.aws-codestarnotifications==1.110.0
aws-cdk.aws-cognito==1.110.0
aws-cdk.aws-ec2==1.110.0
aws-cdk.aws-ecr==1.110.0
aws-cdk.aws-ecr-assets==1.110.0
aws-cdk.aws-ecs==1.110.0
aws-cdk.aws-efs==1.110.0
aws-cdk.aws-eks==1.109.0
aws-cdk.aws-elasticloadbalancing==1.110.0
aws-cdk.aws-elasticloadbalancingv2==1.110.0
aws-cdk.aws-events==1.110.0
aws-cdk.aws-events-targets==1.110.0
aws-cdk.aws-globalaccelerator==1.110.0
aws-cdk.aws-iam==1.110.0
aws-cdk.aws-kinesis==1.110.0
aws-cdk.aws-kinesisfirehose==1.110.0
aws-cdk.aws-kms==1.110.0
aws-cdk.aws-lambda==1.110.0
aws-cdk.aws-logs==1.110.0
aws-cdk.aws-route53==1.110.0
aws-cdk.aws-route53-targets==1.110.0
aws-cdk.aws-s3==1.110.0
aws-cdk.aws-s3-assets==1.110.0
aws-cdk.aws-sam==1.110.0
aws-cdk.aws-secretsmanager==1.110.0
aws-cdk.aws-servicediscovery==1.110.0
aws-cdk.aws-signer==1.110.0
aws-cdk.aws-sns==1.110.0
aws-cdk.aws-sns-subscriptions==1.110.0
aws-cdk.aws-sqs==1.110.0
aws-cdk.aws-ssm==1.110.0
aws-cdk.aws-stepfunctions==1.110.0
aws-cdk.cloud-assembly-schema==1.110.0
aws-cdk.core==1.110.0
aws-cdk.custom-resources==1.110.0
aws-cdk.cx-api==1.110.0
aws-cdk.lambda-layer-awscli==1.109.0
aws-cdk.lambda-layer-kubectl==1.109.0
aws-cdk.region-info==1.110.0
-e git+git@github.com:svmt/aws-svmt-infra.git@182e7a72fa4d2b2bfca7708cbf7c3058db988873#egg=aws_svmt_infra
cattrs==1.5.0
certifi==2020.12.5
constructs==3.3.86
distlib==0.3.1
filelock==3.0.12
jsii==1.30.0
pipenv==2020.11.15
publication==0.0.3
python-dateutil==2.8.1
six==1.16.0
typing-extensions==3.10.0.0
virtualenv==20.4.6
virtualenv-clone==0.5.4
